pyproject.toml
[project]
name = "wwdc-spectra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.2",
    "flow-matching>=1.0.10",
    "huggingface-hub>=1.3.1",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "ipykernel>=7.1.0",
    "lightning>=2.6.0",
    "polymathic-aion>=0.0.2",
    "python-dotenv>=1.2.1",
    "scienceplots>=2.2.0",
    "scikit-learn>=1.8.0",
    "spender>=0.2.9",
    "submitit>=1.5.4",
    "timm>=1.0.24",
    "torch>=2.9.1",
    "wandb>=0.24.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
